requirements.txt
sentence-transformers
faiss-cpu
numpy
pandas
openpyxl
